% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_random_LS.R
\name{generate_random_LS}
\alias{generate_random_LS}
\title{Generate random L-systems object}
\usage{
generate_random_LS(fname, r_seed)
}
\arguments{
\item{fname}{log file name}

\item{r_seed}{seed (also object id)}
}
\value{
Segments of L-Systems at axes origin. Each with segment's coordinates, direction, id, level, function placeholder
}
\description{
Generate a random L-System object from parameters in file.
This is done by generating a random rule from the parameters and then calling generate_LS()
}
