% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_random_LS.R
\name{new_random_LS}
\alias{new_random_LS}
\title{New random L-system objects}
\usage{
new_random_LS(log_name, img_prefix, nimages, a)
}
\arguments{
\item{log_name}{log file name}

\item{img_prefix}{prefix of image file names in batch}

\item{nimages}{number of images in batch}

\item{a}{direction of movement. a single value or a vector to be sampled}
}
\value{
a batch of images in a folder and an updated log file
}
\description{
creating a batch of random LS images, writing them to a folder and writing their parameters into a log file
}
